refactor(test): move INDIRECT tests to tests/success/failure/edge pattern
Converts lookup/indirect.rs → indirect/mod.rs and adds proper
tests/mod.rs, tests/success.rs, tests/failure.rs, tests/edge.rs.
