+++ crates/core/src/eval/functions/lookup/indirect/mod.rs
@@ -63,73 +63,4 @@ fn is_r1c1_ref(s: &str) -> bool {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::indirect_fn;
-    use crate::types::{ErrorKind, Value};
-
-    // --- success ---
-
-    #[test]
-    fn a1_ref_returns_empty() {
-        assert_eq!(
-            indirect_fn(&[Value::Text("A1".to_string())]),
-            Value::Empty
-        );
-    }
-
-    #[test]
-    fn b10_ref_returns_empty() {
-        assert_eq!(
-            indirect_fn(&[Value::Text("B10".to_string())]),
-            Value::Empty
-        );
-    }
-
-    #[test]
-    fn r1c1_ref_returns_empty() {
-        assert_eq!(
-            indirect_fn(&[Value::Text("R1C1".to_string())]),
-            Value::Empty
-        );
-    }
-
-    #[test]
-    fn explicit_a1_true_returns_empty() {
-        assert_eq!(
-            indirect_fn(&[Value::Text("A1".to_string()), Value::Bool(true)]),
-            Value::Empty
-        );
-    }
-
-    // --- edge ---
-
-    #[test]
-    fn empty_string_returns_ref_error() {
-        assert_eq!(
-            indirect_fn(&[Value::Text("".to_string())]),
-            Value::Error(ErrorKind::Ref)
-        );
-    }
-
-    #[test]
-    fn invalid_ref_returns_ref_error() {
-        assert_eq!(
-            indirect_fn(&[Value::Text("InvalidRef!".to_string())]),
-            Value::Error(ErrorKind::Ref)
-        );
-    }
-
-    #[test]
-    fn no_args_returns_na_error() {
-        assert_eq!(indirect_fn(&[]), Value::Error(ErrorKind::NA));
-    }
-
-    #[test]
-    fn a1_false_mode_returns_empty() {
-        // second arg is the a1 flag; false means R1C1 mode, but "A1" is also a valid A1 ref
-        assert_eq!(
-            indirect_fn(&[Value::Text("A1".to_string()), Value::Bool(false)]),
-            Value::Empty
-        );
-    }
-}
+mod tests;
